More testing and fixing fork/lfork.

+++ srcs/corewar/instructions/fork.c
@@ -17,7 +17,7 @@
 **      parent: cycle, registers, live, carry.
 */
 
-static void vm_fork(t_vm *vm, t_process *proc_cur, unsigned int pc)
+static void vm_fork(t_vm *vm, t_process *proc_cur, int pc)
 {
     t_process   *proc;
 
@@ -27,11 +27,11 @@ static void vm_fork(t_vm *vm, t_process *proc_cur, unsigned int pc)
     proc->carry = proc_cur->carry;
 	proc_set_pc(vm, proc, pc);
 	vm->print("Player %d \"%s\" ", proc->champ->id, proc->champ->name);
-	vm->print("forked a new process at PC %u\n", proc->pc);
+	vm->print("forked a new process at PC %d\n", proc->pc);
 	vm->print == &printw ? wait_input() : 0;
 }
 
-static unsigned int	get_target(t_vm *vm)
+static short int	get_target(t_vm *vm)
 {
 	t_arg			*arg;
 
@@ -49,11 +49,11 @@ static unsigned int	get_target(t_vm *vm)
 void	    op_fork(void *vm_ptr)
 {
     t_vm			*vm;
-	unsigned int	val;
+	short int		val;
 
     vm = (t_vm *)vm_ptr;
 	val = get_target(vm);
-	vm->print("fork %u | ", val);
+	vm->print("fork %d | ", val);
     vm_fork(vm, vm->procs.cur, vm->pc + (val % IDX_MOD));
 }
 
@@ -65,10 +65,10 @@ void	    op_fork(void *vm_ptr)
 void	    op_lfork(void *vm_ptr)
 {
     t_vm			*vm;
-	unsigned int	val;
+	short int		val;
 
     vm = (t_vm *)vm_ptr;
 	val = get_target(vm);
-	vm->print("lfork %u | ", val);
+	vm->print("lfork %d | ", val);
     vm_fork(vm, vm->procs.cur, vm->pc + val);
 }
